requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
requests==2.26.0
python-dotenv==0.19.2
nltk==3.6.3
pymupdf==1.19.0
fastapi-cors==0.1.1